chore: move Next-dependent `testing` utils into the `next` subdir
