+++ packages/runtime/src/next/components/tests/makeswift-page-metadata-rendering.test.tsx
@@ -4,7 +4,7 @@ import '@testing-library/jest-dom'
 import { act } from 'react-dom/test-utils'
 import { render } from '@testing-library/react'
 import { ReactRuntime } from '../../../react'
-import * as Testing from '../../../runtimes/react/testing'
+import * as Testing from '../../testing'
 import { type MakeswiftPageSnapshot, type MakeswiftPageDocument } from '../../../client'
 import { type CacheData } from '../../../api/react'
 import { Page as MakeswiftPage } from '../page'
