+++ packages/runtime/src/next/components/tests/prop-controllers/page-table-form-fields-prop-controller.test.tsx
@@ -16,7 +16,7 @@ import { Page } from '../../page'
 import { act } from 'react-dom/test-utils'
 import { ReactRuntime } from '../../../../react'
 import { forwardRef } from 'react'
-import * as Testing from '../../../../runtimes/react/testing'
+import * as Testing from '../../../testing'
 
 describe('Page', () => {
   test('can render TableFormFieldsPropController v0 data', async () => {
